Move db logic to ProjectEntry model
